feat: create user to mongodb
