+++ internal/infrastructure/adapters/repositories/json/entity/database_file.go
@@ -1,6 +1,6 @@
 package entity
 
-type DatabaseFile struct {
-	Users  []*UserEntity  `json:"users"`
-	Spaces []*SpaceEntity `json:"spaces"`
-}
+// type DatabaseFile struct {
+// 	Users  []*UserEntity  `json:"users"`
+// 	Spaces []*SpaceEntity `json:"spaces"`
+// }
