+++ internal/infrastructure/adapters/repositories/json/entity/space_entity.go
@@ -1,13 +1,13 @@
 package entity
 
-import "time"
+// import "time"
 
-type SpaceEntity struct {
-	ID          string    `json:"id"`
-	Name        string    `json:"name"`
-	Description string    `json:"description"`
-	CreatedAt   time.Time `json:"created_at"`
-	CreatedBy   string    `json:"created_by"`
-	UpdatedAt   time.Time `json:"updated_at"`
-	UpdatedBy   string    `json:"updated_by"`
-}
+// type SpaceEntity struct {
+// 	ID          string    `json:"id"`
+// 	Name        string    `json:"name"`
+// 	Description string    `json:"description"`
+// 	CreatedAt   time.Time `json:"created_at"`
+// 	CreatedBy   string    `json:"created_by"`
+// 	UpdatedAt   time.Time `json:"updated_at"`
+// 	UpdatedBy   string    `json:"updated_by"`
+// }
